Move test method to Interceptor Class
